requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-dialogflow-cx==1.24.0
twilio==8.10.0
python-dotenv==1.0.0
google-auth==2.22.0
python-multipart==0.0.6